requirements.txt
psycopg2==2.8.3
aiohttp==3.5.4
async-timeout==3.0.1
asyncio==3.4.3
client==0.0.1
discord.py==1.2.3
oauthlib==3.0.2
pycrypto==2.6.1
pyOpenSSL==16.2.0
requests==2.12.4
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
urllib3==1.19.1
websockets==6.0
yarl==1.3.0
